docs: create Commit Check.yml (#3)
* docs: create Commit Check.yml

This commit adds the initial configuration for a GitHub Actions workflow named 'Commit Check'. The workflow is designed to check the conformity of both Pull Request titles and commit messages to the Conventional Commits standard. This setup ensures that contributions follow a consistent and clear format.

Additionally, the 'Create Commit Check.yml' file serves as a starting point for further customization of the workflow based on specific project requirements. The commit message provides a brief summary of the changes made.
ommit pull requests